% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ate.R
\name{ATE_estim}
\alias{ATE_estim}
\title{Non-parametric ATE estimation from labeled data}
\usage{
ATE_estim(
  statespace,
  partition,
  observed_data,
  treated_partitions,
  pp_rhs_form = "~1",
  windowT = c(0, 1)
)
}
\arguments{
\item{statespace}{Full observation window}

\item{partition}{A spatstat tess object}

\item{observed_data}{Data frame with inferred_process column}

\item{treated_partitions}{Tilenames of treated partitions}

\item{pp_rhs_form}{Formula for intensity (unused, kept for API compat)}

\item{windowT}{Numeric vector c(start, end)}
}
\value{
List with ATE_non_param
}
\description{
Non-parametric ATE estimation from labeled data
}
