% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_algorithm.R
\name{convergence_check}
\alias{convergence_check}
\title{Check convergence of the SEM algorithm}
\usage{
convergence_check(result)
}
\arguments{
\item{result}{Output from adaptive_SEM}
}
\value{
Data frame of proportions per iteration
}
\description{
Plots the proportion of points in each labeling category over iterations.
}
