% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etas.R
\name{.etas_alpha_to_opt}
\alias{.etas_alpha_to_opt}
\title{Map alpha_m from natural scale to unconstrained optimiser coordinate}
\usage{
.etas_alpha_to_opt(alpha, lo, hi)
}
\description{
Map alpha_m from natural scale to unconstrained optimiser coordinate
}
\keyword{internal}
