% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etas.R
\name{.etas_opt_to_alpha}
\alias{.etas_opt_to_alpha}
\title{Map unconstrained optimiser coordinate back to alpha_m}
\usage{
.etas_opt_to_alpha(u, lo, hi)
}
\description{
Map unconstrained optimiser coordinate back to alpha_m
}
\keyword{internal}
