% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelings.R
\name{dumb_labeling}
\alias{dumb_labeling}
\title{Dumb labeling: probabilistic relabeling based on expected counts}
\usage{
dumb_labeling(pp_data, partition, tiles_to_thin, expected_points)
}
\arguments{
\item{pp_data}{Data frame with columns x, y, process, location_process, background}

\item{partition}{A spatstat tess object}

\item{tiles_to_thin}{Integer vector of tile indices to relabel}

\item{expected_points}{Expected number of points per tile}
}
\value{
pp_data with updated inferred_process
}
\description{
Dumb labeling: probabilistic relabeling based on expected counts
}
