% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelings.R
\name{fast_labelling_proposal}
\alias{fast_labelling_proposal}
\title{Fast labeling proposal by random flipping}
\usage{
fast_labelling_proposal(pp_data, partition, partition_process, n_thin)
}
\arguments{
\item{pp_data}{Data frame with inferred_process column}

\item{partition}{A spatstat tess object}

\item{partition_process}{Character vector of process names per tile}

\item{n_thin}{Number of points to flip}
}
\value{
pp_data with updated inferred_process
}
\description{
Fast labeling proposal by random flipping
}
