% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkes.R
\name{generate_hawkes_superposition}
\alias{generate_hawkes_superposition}
\title{Generate superposition of independent Hawkes processes}
\usage{
generate_hawkes_superposition(
  Omega,
  partition,
  time_window,
  partition_processes,
  hawkes_params,
  ...
)
}
\arguments{
\item{Omega}{The full state space}

\item{partition}{A spatstat tess object}

\item{time_window}{Numeric vector c(start, end)}

\item{partition_processes}{Character vector of process names per tile}

\item{hawkes_params}{Named list of parameter lists}

\item{...}{Additional arguments passed to sim_hawkes}
}
\value{
List with superposed data and complete data
}
\description{
Generate superposition of independent Hawkes processes
}
