% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_poisson_data}
\alias{generate_poisson_data}
\title{Generate Poisson point process data under a treatment partition}
\usage{
generate_poisson_data(Omega, partition, treat_prop = 0.5, lambda_1, lambda_2)
}
\arguments{
\item{Omega}{The state space (owin)}

\item{partition}{A spatstat tess object}

\item{treat_prop}{Proportion of tiles to treat}

\item{lambda_1}{Intensity for control process}

\item{lambda_2}{Intensity for treated process}
}
\value{
List with obs_data, partition, and treated indicator
}
\description{
Generate Poisson point process data under a treatment partition
}
