% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkes.R
\name{ks_test_pval}
\alias{ks_test_pval}
\title{KS test p-value for Hawkes process goodness-of-fit}
\usage{
ks_test_pval(realiz, windowT, windowS, hawkes_par, zero_background_region)
}
\arguments{
\item{realiz}{Data frame with columns x, y, t}

\item{windowT}{Numeric vector c(start, end)}

\item{windowS}{An owin object}

\item{hawkes_par}{List with elements mu, alpha, beta, K}

\item{zero_background_region}{An owin object}
}
\value{
Numeric p-value from the KS test
}
\description{
KS test p-value for Hawkes process goodness-of-fit
}
