% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelings.R
\name{local_hawkes_likelihood_ratio_labeling}
\alias{local_hawkes_likelihood_ratio_labeling}
\title{Local likelihood-ratio labeling}
\usage{
local_hawkes_likelihood_ratio_labeling(
  pp_data,
  partition,
  tiles_to_thin,
  statespace,
  windowT,
  hawkes_params,
  n_traversal = 1000,
  sample_size = 10
)
}
\arguments{
\item{pp_data}{Data frame with point process data}

\item{partition}{A spatstat tess object}

\item{tiles_to_thin}{Integer vector of control tile indices}

\item{statespace}{Full observation window}

\item{windowT}{Numeric vector c(start, end)}

\item{hawkes_params}{Hawkes parameter list}

\item{n_traversal}{Number of labeling proposals}

\item{sample_size}{Number of partitions for local likelihood}
}
\value{
List with labellings and likelihoods
}
\description{
Local likelihood-ratio labeling
}
