% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelings.R
\name{naive_labeling}
\alias{naive_labeling}
\title{Naive labeling: use the location-based process labels}
\usage{
naive_labeling(pp_data)
}
\arguments{
\item{pp_data}{Data frame with a \code{location_process} column}
}
\value{
pp_data with \code{inferred_process} set to \code{location_process}
}
\description{
Naive labeling: use the location-based process labels
}
