% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelings.R
\name{oracle_labeling}
\alias{oracle_labeling}
\title{Oracle labeling: use the true process labels}
\usage{
oracle_labeling(pp_data)
}
\arguments{
\item{pp_data}{Data frame with a \code{process} column}
}
\value{
pp_data with \code{inferred_process} set to \code{process}
}
\description{
Oracle labeling: use the true process labels
}
