% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelings.R
\name{plot_flips}
\alias{plot_flips}
\title{Plot flips over iterations for tuning}
\usage{
plot_flips(res)
}
\arguments{
\item{res}{Result from em_style_labelling or adaptive_SEM (adaptive list)}
}
\value{
ggplot object
}
\description{
Plot flips over iterations for tuning
}
