% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_pp}
\alias{plot_pp}
\title{Plot a superposed point process with true labels}
\usage{
plot_pp(obs_data, partition, title = "Point Process Plot")
}
\arguments{
\item{obs_data}{Data frame with columns x, y, t, process, background}

\item{partition}{A spatstat tess object}

\item{title}{Plot title}
}
\value{
Prints a ggplot and returns invisibly
}
\description{
Plot a superposed point process with true labels
}
