% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_pp_inferred}
\alias{plot_pp_inferred}
\title{Plot a point process with inferred labels}
\usage{
plot_pp_inferred(obs_data, partition, title = "Point Process Plot")
}
\arguments{
\item{obs_data}{Data frame with columns x, y, t, process, inferred_process}

\item{partition}{A spatstat tess object}

\item{title}{Plot title}
}
\value{
Prints a ggplot and returns invisibly
}
\description{
Plot a point process with inferred labels
}
