% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelings.R
\name{semi_dumb_labeling}
\alias{semi_dumb_labeling}
\title{Semi-dumb labeling: boundary-weighted probabilistic relabeling}
\usage{
semi_dumb_labeling(
  pp_data,
  partition,
  tiles_to_thin,
  expected_points,
  beta = 1
)
}
\arguments{
\item{pp_data}{Data frame with point process data}

\item{partition}{A spatstat tess object}

\item{tiles_to_thin}{Integer vector of tile indices}

\item{expected_points}{Expected number of points per tile}

\item{beta}{Distance weighting parameter}
}
\value{
pp_data with updated inferred_process
}
\description{
Semi-dumb labeling: boundary-weighted probabilistic relabeling
}
