% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkes.R
\name{sim_hawkes}
\alias{sim_hawkes}
\title{Simulate a spatio-temporal Hawkes process}
\usage{
sim_hawkes(
  params,
  windowT,
  windowS,
  background_realization = NULL,
  filtration = NULL,
  optimized = TRUE,
  covariate_lookup = NULL,
  background_rate_var = NULL
)
}
\arguments{
\item{params}{List with elements mu, alpha, beta, K}

\item{windowT}{Numeric vector c(start, end)}

\item{windowS}{An owin object or numeric vector c(xmin, xmax, ymin, ymax)}

\item{background_realization}{Optional list to use as background events}

\item{filtration}{Optional data frame of history points that can trigger}

\item{optimized}{Logical; use vectorized aftershock generation}

\item{covariate_lookup}{Function(x,y) or Raster for inhomogeneous background}

\item{background_rate_var}{Column name for covariate values}
}
\value{
List with elements x, y, t, n, background, W
}
\description{
Simulate a spatio-temporal Hawkes process
}
