% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelings.R
\name{single_flip_proposals}
\alias{single_flip_proposals}
\title{Generate all single-flip labelling proposals}
\usage{
single_flip_proposals(post, pre)
}
\arguments{
\item{post}{Data frame of post-treatment points (with inferred_process)}

\item{pre}{Data frame of pre-treatment points (all labelled "control")}
}
\value{
List of N data frames, each with one point's label flipped
}
\description{
For each post-treatment point, create one proposal where that point's
label is toggled (control <-> treated). Returns N proposals for N post
points -- exhaustive over all single-point moves.
}
