% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ate.R
\name{tau_i}
\alias{tau_i}
\title{Estimate the one-flip treatment effect for cell i}
\usage{
tau_i(
  i,
  partition,
  treated_partitions,
  statespace,
  windowT,
  control_pp,
  treated_pp,
  n_sim
)
}
\arguments{
\item{i}{Cell index}

\item{partition}{A spatstat tess object}

\item{treated_partitions}{Tilenames of treated partitions}

\item{statespace}{Full observation window}

\item{windowT}{Numeric vector c(start, end)}

\item{control_pp}{Control Hawkes parameters}

\item{treated_pp}{Treated Hawkes parameters}

\item{n_sim}{Number of simulations}
}
\value{
Scalar estimate of tau_i
}
\description{
Simulates under z_plus (cell i treated) and z_minus (cell i control),
returns the difference in expected counts in cell i.
}
