% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ate.R
\name{tau_i_etas}
\alias{tau_i_etas}
\title{Estimate one-flip treatment effect for cell i (ETAS model)}
\usage{
tau_i_etas(
  i,
  partition,
  treated_partitions,
  statespace,
  windowT,
  control_pp,
  treated_pp,
  n_sim,
  m0,
  beta_gr = NULL,
  mag_pool = NULL
)
}
\arguments{
\item{i}{Cell index}

\item{partition}{A spatstat tess object}

\item{treated_partitions}{Tilenames of treated partitions}

\item{statespace}{Full observation window}

\item{windowT}{Numeric vector c(start, end)}

\item{control_pp}{Control ETAS parameter list.}

\item{treated_pp}{Treated ETAS parameter list.}

\item{n_sim}{Number of simulations}

\item{m0}{Reference magnitude.}

\item{beta_gr}{Gutenberg-Richter beta (or NULL for resampling).}

\item{mag_pool}{Magnitude pool for resampling.}
}
\value{
Scalar estimate of tau_i.
}
\description{
Analogue of \code{tau_i} for the ETAS model.
}
