% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{tess_to_sf}
\alias{tess_to_sf}
\title{Convert a spatstat tessellation to an sf POLYGON layer}
\usage{
tess_to_sf(x)
}
\arguments{
\item{x}{A spatstat tess object}
}
\value{
An sf data frame with tile_id and geometry columns
}
\description{
Convert a spatstat tessellation to an sf POLYGON layer
}
